feat: add support for custom error messages
