+++ config.json
@@ -1,24 +1,34 @@
 {
 	"internal_prefix": "/internal",
 	"openapi_boostrap": "openapi.yaml",
-	"logging": {
-    "version": 1,
-    "disable_existing_loggers": false,
-    "formatters": {
-			"default": {
-				"()": "uvicorn.logging.DefaultFormatter",
-				"fmt": "%(levelprefix)s %(message)s",
-				"datefmt": "%Y-%m-%d %H:%M:%S"
+	"settings": {
+		"error_responses": [
+			{
+				"status_code": 404,
+    			"body": {
+					"error": "Page not found"
+				}
 			}
-    },
-    "handlers": {
+		]
+	},
+	"logging": {
+		"version": 1,
+		"disable_existing_loggers": false,
+		"formatters": {
+				"default": {
+					"()": "uvicorn.logging.DefaultFormatter",
+					"fmt": "%(levelprefix)s %(message)s",
+					"datefmt": "%Y-%m-%d %H:%M:%S"
+				}
+		},
+		"handlers": {
 			"default": {
 				"formatter": "default",
 				"class": "logging.StreamHandler",
 				"stream": "ext://sys.stderr"
 			}
 		},
-    "loggers": {
+		"loggers": {
 			"trickster": {
 				"handlers": ["default"],
 				"level": "DEBUG"
