+++ tests/mocked_files/config.json
@@ -1,24 +1,32 @@
 {
 	"internal_prefix": "/internal",
 	"openapi_boostrap": "openapi.yaml",
+	"settings": {
+		"error_response": [
+			{
+				"status_code": 404,
+				"message": "Page not found!"
+			}
+		]
+	},
 	"logging": {
-    "version": 1,
-    "disable_existing_loggers": false,
-    "formatters": {
+		"version": 1,
+		"disable_existing_loggers": false,
+		"formatters": {
 			"default": {
 				"()": "uvicorn.logging.DefaultFormatter",
 				"fmt": "%(levelprefix)s %(message)s",
 				"datefmt": "%Y-%m-%d %H:%M:%S"
 			}
-    },
-    "handlers": {
+		},
+		"handlers": {
 			"default": {
 				"formatter": "default",
 				"class": "logging.StreamHandler",
 				"stream": "ext://sys.stderr"
 			}
 		},
-    "loggers": {
+		"loggers": {
 			"trickster": {
 				"handlers": ["default"],
 				"level": "ERROR"
